+++ lib/bmg/algebra.rb
@@ -123,15 +123,6 @@ module Bmg
     end
     protected :_restrict
 
-    def rxmatch(attrs, matcher, options = {})
-      _rxmatch type.rxmatch(attrs, matcher, options), attrs, matcher, options
-    end
-
-    def _rxmatch(type, attrs, matcher, options)
-      Operator::Rxmatch.new(type, self, attrs, matcher, options)
-    end
-    protected :_rxmatch
-
     def union(other, options = {})
       _union self.type.union(other.type), other, options
     end
@@ -150,5 +141,7 @@ module Bmg
       self
     end
 
+    require_relative 'algebra/shortcuts'
+    include Shortcuts
   end # module Algebra
 end # module Bmg